@@ -1,14 +1,33 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\ProcessedActivity;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Foundation\Testing\WithFaker;
 
-$factory->define(ProcessedActivity::class, function (Faker $faker) {
-    return [
-        'user_id' => factory(User::class),
-        'activity_id' => $faker->randomNumber(),
-    ];
-});
+class ProcessedActivityFactory extends Factory
+{
+    use WithFaker;
+
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = ProcessedActivity::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'user_id' => User::factory(),
+            'activity_id' => $this->faker->randomNumber(),
+        ];
+    }
+}
