@@ -1,17 +1,33 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\SocialToken;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
-$factory->define(SocialToken::class, function (Faker $faker) {
-    return array(
-        'user_id' => factory(User::class),
-        'token' => Str::random(32),
-        'refresh_token' => Str::random(32),
-        'expires_at' => now()->addSeconds(60),
-    );
-});
+class SocialTokenFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = SocialToken::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'user_id' => User::factory(),
+            'token' => Str::random(32),
+            'refresh_token' => Str::random(32),
+            'expires_at' => now()->addSeconds(60),
+        ];
+    }
+}
